LPD-35627 Move PortletConfigurationPortletKeys to the new api module to reuse it
