+++ modules/apps/portlet-configuration/portlet-configuration-icon/portlet-configuration-icon-api/src/main/java/com/liferay/portlet/configuration/icon/constants/PortletConfigurationPortletKeys.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portlet.configuration.web.internal.constants;
+package com.liferay.portlet.configuration.icon.constants;
 
 /**
  * @author Juergen Kappler
