+++ modules/apps/portlet-configuration/portlet-configuration-icon/portlet-configuration-icon-api/src/main/resources/com/liferay/portlet/configuration/icon/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
